% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAgeBatchScaleResults.R
\name{getAgeBatchScaleResults}
\alias{getAgeBatchScaleResults}
\title{Get Age Batch Scale Results}
\usage{
getAgeBatchScaleResults(
  batch_id,
  config_file = "saaWeb.config",
  user_name = Sys.getenv("username"),
  password = NULL
)
}
\arguments{
\item{config_file}{Configuration file}

\item{user_name}{User name to execute the query as}

\item{password}{Password of the user}

\item{batch_ids}{Batch IDs to retrieve scale age results}
}
\value{
Results of the Query
}
\description{
Get Age Batch Scale Results
}
\examples{
getAgeBatchScaleResults(114105)

}
