% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{hasText}
\alias{hasText}
\title{Has Non-Zero Text Value}
\usage{
hasText(text)
}
\arguments{
\item{text}{Text value to check}
}
\description{
Check to see if the parameter has a non-zero text value
}
\examples{
hasText(c(NA, "", "Salmon", "   "))
}
