% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{loadConfigFile}
\alias{loadConfigFile}
\title{Load Configuration File}
\usage{
loadConfigFile(config_file_name)
}
\arguments{
\item{config_file_name}{Configuration file name}
}
\description{
Load Configuration File
}
