% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openSaaWebConnection.R
\name{openSaaWebConnection}
\alias{openSaaWebConnection}
\title{Open SAA Web Connection}
\usage{
openSaaWebConnection(usage_url, user_name, password)
}
\arguments{
\item{usage_url}{URL for network usage acceptance}

\item{user_name}{User name to execute the query as}

\item{password}{Password of the user}
}
\value{
Curl Session Handle
}
\description{
Open SAA Web Connection
}
\examples{
example_query_file <- system.file("query_doc", "tag_release.qry", package = "saaWeb")
runCwtExtractorQuery(example_query_file)

}
