% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAgeBatchScaleResults.R
\name{parseAgeScaleReads}
\alias{parseAgeScaleReads}
\title{Parse Age Scale Reads}
\usage{
parseAgeScaleReads(scale_read_df)
}
\arguments{
\item{scale_read_df}{Data frame of scale age reads}
}
\value{
Parsed and formatted scale age reads
}
\description{
Parse Age Scale Reads
}
