% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExtratorQuery.R
\name{parseExtractorResponse}
\alias{parseExtractorResponse}
\title{Parse Extractor Response}
\usage{
parseExtractorResponse(query_response)
}
\arguments{
\item{query_response}{Query response}
}
\value{
Data frame based on the response
}
\description{
Parse Extractor Response
}
