% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCwtExtractorQuery.r
\name{runCwtExtractorQuery}
\alias{runCwtExtractorQuery}
\title{Run CWT Extractor Query}
\usage{
runCwtExtractorQuery(
  query_doc,
  config_file = "saaWeb.config",
  user_name = Sys.getenv("username"),
  password = NULL
)
}
\arguments{
\item{query_doc}{JSON Query file name or text}

\item{config_file}{Configuration file for SAA Web Services}

\item{user_name}{User name to execute the query as}

\item{password}{Password of the user}
}
\value{
Results of the Query
}
\description{
Run CWT Extractor Query
}
\examples{
example_query_file <- system.file("query_doc", "tag_release.qry", package = "saaWeb")
runCwtExtractorQuery(example_query_file)

}
