% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runNuSedsExtractorQuery.r
\name{runNusedsExtractorQuery}
\alias{runNusedsExtractorQuery}
\title{Run NuSEDs Extractor Query}
\usage{
runNusedsExtractorQuery(
  query_doc,
  config_file = "saaWeb.config",
  user_name = Sys.getenv("username"),
  password = NULL
)
}
\arguments{
\item{query_doc}{JSON Query file name or text}

\item{config_file}{Configuration file for SAA Web Services}

\item{user_name}{User name to execute the query as}

\item{password}{Password of the user}
}
\value{
Results of the Query
}
\description{
Run NuSEDs Extractor Query
}
