% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAgeBatchDetails.R
\name{getAgeBatchDetails}
\alias{getAgeBatchDetails}
\title{Get Age Batch List}
\usage{
getAgeBatchDetails(
  batch_id,
  config_file = "saaWeb.config",
  user_name = Sys.getenv("username"),
  password = NULL
)
}
\arguments{
\item{config_file}{Configuration file}

\item{user_name}{User name to execute the query as}

\item{password}{Password of the user}

\item{years}{Years to retrieve}
}
\value{
Results of the Query
}
\description{
Get Age Batch List
}
\examples{
AgeBatchDetail(114105)

}
