% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAgeBatchList.R
\name{getAgeBatchList}
\alias{getAgeBatchList}
\title{Get Age Batch List}
\usage{
getAgeBatchList(
  config_file = "saaWeb.config",
  user_name = Sys.getenv("username"),
  password = NULL
)
}
\arguments{
\item{config_file}{Configuration file}

\item{user_name}{User name to execute the query as}

\item{password}{Password of the user}
}
\value{
Results of the Query
}
\description{
Retrieve header information on all age batches
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
getAgeBatchList()
\dontshow{\}) # examplesIf}
}
