% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAgeBatchScaleResults.R
\name{getAgeBatchScaleResults}
\alias{getAgeBatchScaleResults}
\title{Get Age Batch Scale Results}
\usage{
getAgeBatchScaleResults(
  batch_ids,
  config_file = "saaWeb.config",
  user_name = Sys.getenv("username"),
  password = NULL
)
}
\arguments{
\item{batch_ids}{Batch IDs to retrieve scale age results}

\item{config_file}{Configuration file}

\item{user_name}{User name to execute the query as}

\item{password}{Password of the user}
}
\value{
Results of the Query
}
\description{
Retrieve the scale age results for the specific batches
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
getAgeBatchScaleResults(114105)
\dontshow{\}) # examplesIf}
}
