% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExtratorQuery.r
\name{parseExtractorResponse}
\alias{parseExtractorResponse}
\title{Parse Extractor Response}
\usage{
parseExtractorResponse(query_reponse)
}
\arguments{
\item{query_reponse}{Query response}
}
\value{
Data frame based on the response
}
\description{
Parse Extractor Response
}
