% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExtratorQuery.r
\name{runExtractorQuery}
\alias{runExtractorQuery}
\title{Run Extractor Query}
\usage{
runExtractorQuery(
  query_doc,
  user_name = Sys.getenv("username"),
  password = NULL
)
}
\arguments{
\item{query_doc}{JSON Query file name or text}

\item{user_name}{User name to execute the query as}

\item{password}{Password of the user}
}
\value{
Results of the Query
}
\description{
Run Extractor Query
}
\examples{
example_query_file <- system.file("query_doc", "tag_release.qry", package = "saaWeb")
runExtractorQuery(example_query_file)

}
